pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "odoo-mcp"
version = "0.1.0"
description = "A Model Context Protocol server for Odoo"
requires-python = ">=3.12"
license = { text = "MPL-2.0" }
dependencies = [
    "mcp>=1.26.0,<2",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.27.0",
    "cachetools>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "ty>=0.0.1",
    "ruff>=0.11.12",
    "requests>=2.32.3",
]

[project.scripts]
odoo-mcp = "odoo_mcp.__main__:main"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["odoo_mcp"]
omit = ["tests/*", "*/__pycache__/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["odoo_mcp/"]
exclude = ["venv/", ".venv/", "build/", "dist/"]

[tool.ty.rules]
possibly-unresolved-reference = "warn"
invalid-argument-type = "ignore"
invalid-parameter-default = "error"
invalid-return-type = "warn"
invalid-assignment = "warn"
unresolved-attribute = "ignore"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "Q"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "ty>=0.0.1",
]
